+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - fix(hashes): Fix invoke tx hashes for version v3
+- fix(program): Changed visibility of program serializer + archived
 - fix(classes): remove the classes ordenation to allow fork rebasements
 - fix(felt): enforce Felt type as much as possible into compute_hash.rs
 - fix(logs): fixed some logs and others